pyproject.toml
[project]
name = "scanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "nanonispy2>=1.2.0",
    "numpy>=2.3.5",
    "pyqt6>=6.10.1",
    "pyqtgraph>=0.14.0",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
]
